Create a separate OrderSubtotal view to display before discounts and taxes
See #8578
