+++ assets/css/admin/style.scss
@@ -669,6 +669,7 @@ table.edd-order-overview-summary--refund {
 }
 
 /** Totals */
+.edd-order-overview-summary__subtotal .column-primary,
 .edd-order-overview-summary__totals .column-primary {
 	font-weight: 600;
 }
