+++ assets/js/admin/orders/order-overview/views/summary.js
@@ -4,6 +4,7 @@
  * Internal dependencies
  */
 import { OrderItems } from './order-items.js';
+import { OrderSubtotal } from './order-subtotal.js';
 import { OrderAdjustments } from './order-adjustments.js';
 import { Refunds } from './order-refunds.js';
 import { Totals } from './totals.js';
@@ -33,7 +34,9 @@ export const Summary = wp.Backbone.View.extend( {
 	 */
 	render() {
 		this.views.add( new OrderItems( this.options ) );
+		this.views.add( new OrderSubtotal( this.options ) );
 		this.views.add( new OrderAdjustments( this.options ) );
+		// @todo Tax output should be a part of OrderAdjustments.
 		this.views.add( new Totals( this.options ) );
 		this.views.add( new Refunds( this.options ) );
