+++ includes/admin/views/tmpl-order-totals.php
@@ -10,16 +10,6 @@
  */
 ?>
 
-<tr class="is-expanded">
-	<td></td>
-	<td colspan="{{ data.config.colspan }}" class="column-primary">
-		<?php esc_html_e( 'Subtotal', 'easy-digital-downloads' ); ?>
-	</td>
-	<td class="column-right" data-colname="<?php esc_html_e( 'Amount', 'easy-digital-downloads' ); ?>">
-		{{ data.subtotalCurrency }}
-	</td>
-</tr>
-
 <# if ( 'none' !== data.state.hasTax ) { #>
 
 	<tr class="is-expanded">
@@ -93,7 +83,6 @@
 	</td>
 </tr>
 
-<input type="hidden" value="{{ data.subtotal }}" name="subtotal" />
 <input type="hidden" value="{{ data.tax }}" name="tax" />
 <input type="hidden" value="{{ data.discount }}" name="discount" />
 <input type="hidden" value="{{ data.total }}" name="total" />
